requirements.txt
build==0.5.1
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
packaging==21.0
pep517==0.11.0
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
toml==0.10.2
tomli==1.1.0
urllib3==1.26.6
